backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
Pillow==7.0.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.1
virtualenv==20.0.13
wrapt==1.11.2
zipp==3.1.0